requirements.txt
websocket-client
ollama
opencv-python
Pillow
requests
sv_ttk
